Refactor: Separate logging configuration into its own module
I've moved the logging setup from `src/app/main.py` to a new dedicated module `src/app/core/logging_config.py`.

This change helps to declutter `main.py` and centralizes the logging configuration, making it easier for you to manage and modify in the future.

The `main.py` file now imports the necessary logger (`app_logger` for its side effect of configuring the root 'app' logger) from the new module and defines its own `app.main` logger which inherits the established configuration.
